Reorganized common extensions. Words count helpers are converted to extensions too.
