+++ Warehouse.Application/Warehouse.Application.csproj
@@ -12,7 +12,6 @@
   </ItemGroup>
   <ItemGroup>
     <None Remove="DTOs\" />
-    <None Remove="Helpers\" />
     <None Remove="Products\" />
     <None Remove="Interfaces\" />
     <None Remove="Interfaces\Repositories\" />
@@ -22,7 +21,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="DTOs\" />
-    <Folder Include="Helpers\" />
     <Folder Include="Products\" />
     <Folder Include="Interfaces\" />
     <Folder Include="Interfaces\Repositories\" />
